week6/Day1/extension/cursorclone/requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.15
aiosignal==1.4.0
annotated-types==0.7.0
anyio==4.10.0
async-timeout==4.0.3
attrs==25.3.0
certifi==2025.8.3
charset-normalizer==3.4.2
colorama==0.4.6
dataclasses-json==0.6.7
distro==1.9.0
exceptiongroup==1.3.0
frozenlist==1.7.0
greenlet==3.2.3
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
httpx-sse==0.4.1
idna==3.10
jiter==0.10.0
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.27
langchain-community==0.3.27
langchain-core==0.3.72
langchain-openai==0.3.28
langchain-text-splitters==0.3.9
langgraph==0.6.3
langgraph-checkpoint==2.1.1
langgraph-prebuilt==0.6.3
langgraph-sdk==0.2.0
langsmith==0.4.12
marshmallow==3.26.1
multidict==6.6.3
mypy-extensions==1.1.0
numpy==2.2.6
openai==1.99.1
orjson==3.11.1
ormsgpack==1.10.0
packaging==25.0
pip==21.2.4
propcache==0.3.2
pydantic==2.11.7
pydantic-core==2.33.2
pydantic-settings==2.10.1
python-dotenv==1.1.1
pyyaml==6.0.2
regex==2025.7.34
requests==2.32.4
requests-toolbelt==1.0.0
setuptools==58.1.0
sniffio==1.3.1
sqlalchemy==2.0.42
tavily-python==0.7.10
tenacity==9.1.2
tiktoken==0.10.0
tqdm==4.67.1
typing-extensions==4.14.1
typing-inspect==0.9.0
typing-inspection==0.4.1
urllib3==2.5.0
xxhash==3.5.0
yarl==1.20.1
zstandard==0.23.0
